+++ Plutarch/Builtin/Data.hs
@@ -17,8 +17,6 @@ module Plutarch.Builtin.Data (
   pserialiseData,
   pconstrBuiltin,
   PBuiltinPair (PBuiltinPair),
-  pfstBuiltin,
-  psndBuiltin,
   ppairDataBuiltin,
   PBuiltinList (PCons, PNil),
   pheadBuiltin,
@@ -86,14 +84,6 @@ pconstrBuiltin = punsafeBuiltin PLC.ConstrData
 -}
 data PBuiltinPair (a :: S -> Type) (b :: S -> Type) (s :: S)
   = PBuiltinPair (Term s a) (Term s b)
-
-{-# DEPRECATED pfstBuiltin "Use pmatch instead" #-}
-pfstBuiltin :: Term s (PBuiltinPair a b :--> a)
-pfstBuiltin = phoistAcyclic $ pforce . pforce . punsafeBuiltin $ PLC.FstPair
-
-{-# DEPRECATED psndBuiltin "Use pmatch instead" #-}
-psndBuiltin :: Term s (PBuiltinPair a b :--> b)
-psndBuiltin = phoistAcyclic $ pforce . pforce . punsafeBuiltin $ PLC.SndPair
 
 {- | Construct a builtin pair of 'PData' elements.