+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ## Removed
 
 * `pif'`, as there's no reason to use it any more
+* `pfstBuiltin` and `psndBuiltin`, as `pmatch` is more efficient
 
 # 1.13.0 -- 04-03-2026